Reset sse-starlette's cached exit event around every test
sse-starlette below 2.0 (the lowest supported version) keeps a
module-level event on AppStatus that becomes bound to the first event
loop that waits on it, so a test serving an SSE response in process can
inherit an event created on another test's loop and fail. The per-test
reset that test_sse_security.py already carried moves to tests/conftest.py
so the in-process Streamable HTTP tests get it too.
